Forcing a window to the foreground is again done by AttachThreadInput.
Initial work to using AppBars for the bars. This is going to fix a host of problems. The only caveat being that the Windows Taskbar must be on autohide for an AppBar to be fixed correctly to the same side as the Taskbar.

--HG--
extra : convert_revision : svn%3A4bd69a10-c447-994a-b4ab-e5eff1cba757/trunk%4061